requirements.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2
boto3
django-environ
psycopg2-binary
djangorestframework
